feature: 标签管理功能 #178

+++ src/backend/job-manage/service-job-manage/src/main/java/com/tencent/bk/job/manage/service/impl/TagServiceImpl.java
@@ -272,12 +272,16 @@ public class TagServiceImpl implements TagService {
     public void batchPatchResourceTags(List<ResourceTagDTO> addResourceTags,
                                           List<ResourceTagDTO> deleteResourceTags) {
         StopWatch watch = new StopWatch("batchPatchResourceTags");
-        watch.start("batchSaveResourceTags");
-        resourceTagDAO.batchSaveResourceTags(addResourceTags);
-        watch.stop();
-        watch.start("deleteResourceTags");
-        resourceTagDAO.deleteResourceTags(deleteResourceTags);
-        watch.stop();
+        if (CollectionUtils.isNotEmpty(addResourceTags)) {
+            watch.start("batchSaveResourceTags");
+            resourceTagDAO.batchSaveResourceTags(addResourceTags);
+            watch.stop();
+        }
+        if (CollectionUtils.isNotEmpty(deleteResourceTags)) {
+            watch.start("deleteResourceTags");
+            resourceTagDAO.deleteResourceTags(deleteResourceTags);
+            watch.stop();
+        }
         if (watch.getTotalTimeMillis() > 1000L) {
             log.warn("BatchPatchResourceTags cost too much time, cost: {} ms, statistics: {}",
                 watch.getTotalTimeMillis(), watch.prettyPrint());
